backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# AI/ML Stack - Fixed versions
langchain==0.1.16
langchain-community==0.0.38
langgraph==0.0.43
ollama==0.1.7
chromadb==0.4.24
sentence-transformers==2.5.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
redis==5.0.1

# Infrastructure Tools
python-terraform==0.10.1
boto3==1.34.51
kubernetes==30.1.0
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
google-cloud-resource-manager==1.12.0

# Integration
jira==3.6.0
gitpython==3.1.42
python-jenkins==1.8.2
requests==2.31.0

# Security & Policy
cryptography==42.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1
pyyaml==6.0.1
aiofiles==23.2.1
httpx>=0.25.2

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5

# Monitoring
prometheus-client==0.20.0
mlflow-skinny==2.17.2
